Refactored all components into files
